backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-gedcom==1.0.0
python-multipart==0.0.9
pytest==7.4.3
httpx==0.25.2
minio==7.2.0
python-docx==1.1.0
pdfplumber==0.10.3
Pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0